creating predefined examinations as a clinic admin

+++ src/main/java/com/clinicCenter/service/implementation/MedicalExaminationServiceImpl.java
@@ -44,7 +44,7 @@ public class MedicalExaminationServiceImpl implements MedicalExaminationService 
         User patient = userRepository.findById(patientId).get();
         System.out.println("preuzet pacijent id : " + patient.getId());
 
-        MedicalExaminationRequest newReq = new MedicalExaminationRequest(type, date, clinic, (Doctor)doctor, (Patient)patient);
+        MedicalExaminationRequest newReq = new MedicalExaminationRequest(type, date, clinic, (Doctor) doctor, (Patient) patient);
         medicalExaminationRequestRepository.save(newReq);
     }
 
@@ -59,7 +59,7 @@ public class MedicalExaminationServiceImpl implements MedicalExaminationService 
     }
 
     @Override
-    public void saveExamination(Date date, Double price, Double duration, Double discount, Long roomId, Long clinicId, Long doctorId, Long patientId, Long typeId, Long requestId) {
+    public void saveExamination(Date date, Double price, Double duration, Double discount, Long roomId, Long clinicId, Long doctorId, Long patientId, Long typeId, Long requestId, Boolean predefined) {
 
         MedicalExaminationType type = medicalExaminationTypeRepository.findById(typeId).get();
         System.out.println("preuzet tip id : " + type.getId());
@@ -67,39 +67,47 @@ public class MedicalExaminationServiceImpl implements MedicalExaminationService 
         System.out.println("preuzet clinic id : " + clinic.getId());
         User doctor = userRepository.findById(doctorId).get();
         System.out.println("preuzet doktor id : " + doctor.getId());
-        User patient = userRepository.findById(patientId).get();
-        System.out.println("preuzet pacijent id : " + patient.getId());
+
+        User patient = null;
+        if (patientId != null) {
+            patient = userRepository.findById(patientId).get();
+            System.out.println("preuzet pacijent id : " + patient.getId());
+        }
+
         MedicalExaminationRoom room = medicalExaminationRoomRepository.findById(roomId).get();
         System.out.println("preuzet room ud : " + room.getId());
 
-        MedicalExamination newExam = new MedicalExamination(date, price, duration, discount, room, (Doctor)doctor, (Patient)patient, clinic, type);
+        MedicalExamination newExam = new MedicalExamination(date, price, duration, discount, room, (Doctor) doctor, (Patient) patient, clinic, type, predefined);
         this.medicalExaminationRepository.save(newExam);
 
-        this.medicalExaminationRequestRepository.deleteById(requestId);
+        if (requestId != null) {
+            this.medicalExaminationRequestRepository.deleteById(requestId);
+            String message = "You have scheduled an examination : " +
+                    "\n Date : " + newExam.getDate() +
+                    "\n Clinic : " + newExam.getClinic().getName() + " , " + newExam.getClinic().getAddress() + " , " + newExam.getClinic().getCity() +
+                    "\n Doctor : " + newExam.getDoctor().getFirstName() + " " + newExam.getDoctor().getLastName() +
+                    "\n Examination type : " + newExam.getType().getName() +
+                    "\n Examination room : " + newExam.getMedicalExaminationRoom().getName() + " " + newExam.getMedicalExaminationRoom().getNumber() +
+                    "\n Price : " + newExam.getPrice() +
+                    "\n Discount : " + newExam.getDiscount() +
+                    "\n Duration : " + newExam.getDuration() +
+                    "\n  " +
+                    "\n Confirm : " + "http://localhost:4200/confirmScheduledExamination/" + newExam.getId() +
+                    "\n Decline : " + "http://localhost:4200/declineScheduledExamination/" + newExam.getId();
 
-        String message = "You have scheduled an examination : " +
-                "\n Date : " + newExam.getDate() +
-                "\n Clinic : " + newExam.getClinic().getName() + " , " + newExam.getClinic().getAddress() + " , " + newExam.getClinic().getCity() +
-                "\n Doctor : " + newExam.getDoctor().getFirstName() + " " + newExam.getDoctor().getLastName() +
-                "\n Examination type : " + newExam.getType().getName() +
-                "\n Examination room : " + newExam.getMedicalExaminationRoom().getName() + " " + newExam.getMedicalExaminationRoom().getNumber() +
-                "\n Price : " + newExam.getPrice() +
-                "\n Discount : " + newExam.getDiscount() +
-                "\n Duration : " + newExam.getDuration() +
-                "\n  " +
-                "\n Confirm : " + "http://localhost:4200/confirmScheduledExamination/" + newExam.getId() +
-                "\n Decline : " + "http://localhost:4200/declineScheduledExamination/" + newExam.getId();
+            String message2 = "You have an examination scheduled : " +
+                    "\n Date : " + newExam.getDate() +
+                    "\n Clinic : " + newExam.getClinic().getName() + " , " + newExam.getClinic().getAddress() + " , " + newExam.getClinic().getCity() +
+                    "\n Patient : " + newExam.getPatient().getFirstName() + " " + newExam.getPatient().getLastName() +
+                    "\n Examination type : " + newExam.getType().getName() +
+                    "\n Examination room : " + newExam.getMedicalExaminationRoom().getName() + " " + newExam.getMedicalExaminationRoom().getNumber() +
+                    "\n Duration : " + newExam.getDuration();
 
-        String message2 = "You have an examination scheduled : " +
-                "\n Date : " + newExam.getDate() +
-                "\n Clinic : " + newExam.getClinic().getName() + " , " + newExam.getClinic().getAddress() + " , " + newExam.getClinic().getCity() +
-                "\n Patient : " + newExam.getPatient().getFirstName() + " " + newExam.getPatient().getLastName() +
-                "\n Examination type : " + newExam.getType().getName() +
-                "\n Examination room : " + newExam.getMedicalExaminationRoom().getName() + " " + newExam.getMedicalExaminationRoom().getNumber() +
-                "\n Duration : " + newExam.getDuration();
+            emailController.sendMail(patient.getEmail(), message, "Automated mail : Confirm or decline scheduled examination");
+            emailController.sendMail(doctor.getEmail(), message2, "Automated mail : Confirm or decline scheduled examination");
 
-        emailController.sendMail(patient.getEmail(), message, "Automated mail : Confirm or decline scheduled examination");
-        emailController.sendMail(doctor.getEmail(), message2, "Automated mail : Confirm or decline scheduled examination");
+        }
+
     }
 
     @Override
